+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add rudimentary way of differentiating between different automation types
 * Disable "heating/cooling allowed" on legacy EDA units, fixes crash (https://github.com/Jalle19/eda-modbus-bridge/issues/105)
 * Disable "eco" mode switch on older units (https://github.com/Jalle19/eda-modbus-bridge/issues/104)
+* Add optional sensors for control panel temperature, supply fan speed and exhaust fan speed (https://github.com/Jalle19/eda-modbus-bridge/issues/96)
 
 ## 2.7.1